// bindings/typescript/generated/robot_bus_interface/grpc/v1/action_gateway.ts
export interface GoalCommand {
    /**
     * @generated from protobuf field: string action_name = 1
     */
    actionName: string;
    /**
     * @generated from protobuf field: bytes goal = 2
     */
    goal: Uint8Array;
    /**
     * Optional; empty → gateway generates a UUID.
     *
     * @generated from protobuf field: string goal_id = 3
     */
    goalId: string;
    /**
     * Applied to each ZMQ recv while this session is active; 0 → wait forever
     * (session still polls with a short internal tick so CANCEL can arrive).
     *
     * @generated from protobuf field: uint32 timeout_ms = 4
     */
    timeoutMs: number;
}
// ...
export interface ActionEvent {
    /**
     * @generated from protobuf field: string action_name = 1
     */
    actionName: string;
    /**
     * @generated from protobuf field: string goal_id = 2
     */
    goalId: string;
    /**
     * @generated from protobuf field: robot_bus_interface.grpc.v1.ActionKind kind = 3
     */
    kind: ActionKind;
    /**
     * @generated from protobuf field: bytes body = 4
     */
    body: Uint8Array;
}
/**
 * @generated from protobuf enum robot_bus_interface.grpc.v1.ActionKind
 */
export enum ActionKind {
    /**
     * @generated from protobuf enum value: ACTION_KIND_UNSPECIFIED = 0;
     */
    UNSPECIFIED = 0,
    /**
     * @generated from protobuf enum value: ACTION_KIND_GOAL = 1;
     */
    GOAL = 1,
    /**
     * @generated from protobuf enum value: ACTION_KIND_FEEDBACK = 2;
     */
    FEEDBACK = 2,
    /**
     * @generated from protobuf enum value: ACTION_KIND_RESULT = 3;
     */
    RESULT = 3,
    /**
     * @generated from protobuf enum value: ACTION_KIND_CANCEL = 4;
     */
    CANCEL = 4
}
// ...
class GoalCommand$Type extends MessageType<GoalCommand> {
    constructor() {
        super("robot_bus_interface.grpc.v1.GoalCommand", [
            { no: 1, name: "action_name", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "goal", kind: "scalar", T: 12 /*ScalarType.BYTES*/ },
            { no: 3, name: "goal_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 4, name: "timeout_ms", kind: "scalar", T: 13 /*ScalarType.UINT32*/ }
        ]);
    }
    create(value?: PartialMessage<GoalCommand>): GoalCommand {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.actionName = "";
        message.goal = new Uint8Array(0);
        message.goalId = "";
        message.timeoutMs = 0;
        if (value !== undefined)
            reflectionMergePartial<GoalCommand>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: GoalCommand): GoalCommand {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string action_name */ 1:
                    message.actionName = reader.string();
                    break;
                case /* bytes goal */ 2:
                    message.goal = reader.bytes();
                    break;
                case /* string goal_id */ 3:
                    message.goalId = reader.string();
                    break;
                case /* uint32 timeout_ms */ 4:
                    message.timeoutMs = reader.uint32();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: GoalCommand, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string action_name = 1; */
        if (message.actionName !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.actionName);
        /* bytes goal = 2; */
        if (message.goal.length)
            writer.tag(2, WireType.LengthDelimited).bytes(message.goal);
        /* string goal_id = 3; */
        if (message.goalId !== "")
            writer.tag(3, WireType.LengthDelimited).string(message.goalId);
        /* uint32 timeout_ms = 4; */
        if (message.timeoutMs !== 0)
            writer.tag(4, WireType.Varint).uint32(message.timeoutMs);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message robot_bus_interface.grpc.v1.GoalCommand
 */
export const GoalCommand = new GoalCommand$Type();
// ...
class ActionEvent$Type extends MessageType<ActionEvent> {
    constructor() {
        super("robot_bus_interface.grpc.v1.ActionEvent", [
            { no: 1, name: "action_name", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "goal_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "kind", kind: "enum", T: () => ["robot_bus_interface.grpc.v1.ActionKind", ActionKind, "ACTION_KIND_"] },
            { no: 4, name: "body", kind: "scalar", T: 12 /*ScalarType.BYTES*/ }
        ]);
    }
    create(value?: PartialMessage<ActionEvent>): ActionEvent {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.actionName = "";
        message.goalId = "";
        message.kind = 0;
        message.body = new Uint8Array(0);
        if (value !== undefined)
            reflectionMergePartial<ActionEvent>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ActionEvent): ActionEvent {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string action_name */ 1:
                    message.actionName = reader.string();
                    break;
                case /* string goal_id */ 2:
                    message.goalId = reader.string();
                    break;
                case /* robot_bus_interface.grpc.v1.ActionKind kind */ 3:
                    message.kind = reader.int32();
                    break;
                case /* bytes body */ 4:
                    message.body = reader.bytes();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: ActionEvent, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string action_name = 1; */
        if (message.actionName !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.actionName);
        /* string goal_id = 2; */
        if (message.goalId !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.goalId);
        /* robot_bus_interface.grpc.v1.ActionKind kind = 3; */
        if (message.kind !== 0)
            writer.tag(3, WireType.Varint).int32(message.kind);
        /* bytes body = 4; */
        if (message.body.length)
            writer.tag(4, WireType.LengthDelimited).bytes(message.body);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message robot_bus_interface.grpc.v1.ActionEvent
 */
export const ActionEvent = new ActionEvent$Type();